Add dummy XXX_GET_DEV_HANDLE and XXX_GET_OFFSET macros for the interfaces.
Users can use them in OpenCL code, and they will not have to edit it once we
implement filters for complex interfaces.

Not doing this for the multiformat interface: there will never be generic
filters for this interface, since we do not know the layout of the data
manipulated by this interface.