Modifications of internal handling of tags in StarPU-MPI.
A unique tag called STARPU_MPI_TAG is now used for all
StarPU-MPI communications.

Any communication now works with a envelope + data system.
The envelope contains information about the tag of the data
which will be sent, so as the receiver can get the matching
request from the pending receive requests' hashmap, then 
submit it.