configure: fix OpenMP compiler detection
Detecting if -fopenmp option is supported is not enough to make sure the
given compiler actually supports OpenMP. Firstly because, this option is
not standard (ICC uses -openmp instead), and secondly because omp.h is
not tied to that option and this results in compilation errors with Clang.

Using the generic compiler characteristics AC_OPENMP allows to make sure
OpenMP is correctly supported by the compiler. This might also fix
compilation with ICC as it doesn't support -fopenmp.

Tested with GCC 5.2 and Clang 3.7.