tools: add a python script which parses the trace.rec file
This script can be used to parse the generated trace.rec to report
statistics about the number of events/tasks and their durations. The
output is similar to the starpu_paje_state_stats.in, except that this
one doesn't need a Paje trace and it doesn't depend on R and pj_dump
(from the pajeng repository). This starpu_trace_state_stats.py is also
4x much faster than the starpu_paje_state_stats.in script.

As for paje.trace, STARPU_GENERATE_TRACE=1 must be set to generate the
trace.rec file.

No manpages are installed for this tool.