Rework status change
With pipelining, we may actually be e.g. at the same time executing a task,
waiting for data for the next task, and idle because there will not be more
tasks to pipeline.

So rather use a bitset. This will also make the online profiling feedback
simpler in a further commit.